A bug is lurking ... work in progress ffo-19 is exposing it .... but it is difficult to be understood ....

+++ c/src/sort_utils.h
@@ -186,7 +186,7 @@ sort_utils_insertionsort_dsc_i64 (int64_t *const a,
 
 extern void
 sort_utils_insertionsort_asc_p (void **const a,
-                                const int count);
+                                const size_t count);
